SystemUI - fix Kill App
In rare circumstances, the kill App code of extensible key button
could kill the last app on the stack (the launcher) and cause a hot
reboot.  This keeps the default launcher from being killed and thus
preventing the reboot.

I used some of the code found in our commit here:
http://gerrit.sudoservers.com:8080/#/c/3090/6

+++ packages/SystemUI/src/com/android/systemui/statusbar/policy/ExtensibleKeyButtonView.java
@@ -8,6 +8,7 @@ import android.content.ActivityNotFoundException;
 import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.ResolveInfo;
 import android.content.ServiceConnection;
 import android.hardware.input.InputManager;
 import android.os.Handler;
@@ -24,6 +25,7 @@ import android.view.KeyCharacterMap;
 import android.view.KeyEvent;
 import android.view.View;
 import android.view.ViewConfiguration;
+import android.widget.Toast;
 
 import com.android.internal.statusbar.IStatusBarService;
 import com.android.systemui.statusbar.phone.NavigationBarView;
@@ -134,8 +136,18 @@ public class ExtensibleKeyButtonView extends KeyButtonView {
 
     Runnable mKillTask = new Runnable() {
         public void run() {
+            final Intent intent = new Intent(Intent.ACTION_MAIN);
+            String defaultHomePackage = "com.android.launcher";
+            intent.addCategory(Intent.CATEGORY_HOME);
+            final ResolveInfo res = mContext.getPackageManager().resolveActivity(intent, 0);
+            if (res.activityInfo != null && !res.activityInfo.packageName.equals("android")) {
+                defaultHomePackage = res.activityInfo.packageName;
+            }
             String packageName = mActivityManager.getRunningTasks(1).get(0).topActivity.getPackageName();
-            mActivityManager.forceStopPackage(packageName);
+            if (!defaultHomePackage.equals(packageName)) {
+                    mActivityManager.forceStopPackage(packageName);
+                    Toast.makeText(mContext, R.string.app_killed_message, Toast.LENGTH_SHORT).show();
+            }
         }
     };
